build(reporter): Move FreeMarker and AsciiDoc reporters to own projects
While the AsciiDoc reporter depends on the FreeMarker reporter, it makes
sense to put them into separate plugin projects due to the rather large
additional JRuby dependeny of the AsciiDoc reporter. Migrate both
reporters in one go to avoid an intermediate state where a plugin project
would depend on a reporter that is still in the analyzer.
