+++ plugins/reporters/asciidoc/src/funTest/assets/asciidoc-template-reporter-expected-result.adoc
