+++ plugins/reporters/asciidoc/src/funTest/assets/docbook-template-reporter-expected-result.xml
