+++ plugins/reporters/asciidoc/src/funTest/assets/html-template-reporter-expected-result.html
