+++ plugins/reporters/asciidoc/src/funTest/assets/manpage-template-reporter-expected-result.1
