+++ plugins/reporters/asciidoc/src/main/resources/META-INF/services/org.ossreviewtoolkit.reporter.Reporter
@@ -1,0 +1,4 @@
+org.ossreviewtoolkit.plugins.reporters.asciidoc.DocBookTemplateReporter
+org.ossreviewtoolkit.plugins.reporters.asciidoc.HtmlTemplateReporter
+org.ossreviewtoolkit.plugins.reporters.asciidoc.ManPageTemplateReporter
+org.ossreviewtoolkit.plugins.reporters.asciidoc.PdfTemplateReporter
