+++ plugins/reporters/freemarker/src/funTest/assets/plain-text-template-reporter-expected-results
