+++ plugins/reporters/freemarker/src/funTest/kotlin/PlainTextTemplateReporterFunTest.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.reporter.reporters.freemarker
+package org.ossreviewtoolkit.plugins.reporters.freemarker
 
 import io.kotest.core.TestConfiguration
 import io.kotest.core.spec.style.WordSpec
