+++ plugins/reporters/freemarker/src/main/resources/META-INF/services/org.ossreviewtoolkit.reporter.Reporter
@@ -1,0 +1,1 @@
+org.ossreviewtoolkit.plugins.reporters.freemarker.PlainTextTemplateReporter
