+++ plugins/reporters/freemarker/src/test/kotlin/FreeMarkerTemplateProcessorTest.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.reporter.reporters.freemarker
+package org.ossreviewtoolkit.plugins.reporters.freemarker
 
 import io.kotest.core.spec.style.WordSpec
 import io.kotest.inspectors.forAll
